+++ dev/run_dev.R
@@ -2,7 +2,6 @@ golem::detach_all_attached()
 golem::document_and_reload()
 options(
   golem.app.prod = FALSE,
-  shiny.port = httpuv::randomPort(),
-  shiny.maxRequestSize = 5000 * 1024^2
+  shiny.port = httpuv::randomPort()
 )
 run_app()